use an ignorefile to prevent unwanted email addresses being added
The ignorefile is a file of regular expressions (including plain
strings).  For every email address being added, it is checked against
all the regular expressions.  If it matches any then the email address
is skipped.

The ignorefile is set by the ignorefile value in the notmuch config file
in the [addressbook] section, if present.